moved ui_Versa3dMainWindow.py to GUI folder
